+++ test/test-browser/01_websockets_embedded_sanity_encryptedpayloads.js
@@ -1,75 +1,33 @@
 describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
 
-  var expect = require('expect.js');
-  var happn = require('../lib/index')
-  var service = happn.service;
-  var happn_client = happn.client;
-  var async = require('async');
-
-  var test_secret = 'test_secret';
-  var mode = "embedded";
-  var default_timeout = 4000;
-  var happnInstance = null;
-
-  /*
-  This test demonstrates starting up the happn service - 
-  the authentication service will use authTokenSecret to encrypt web tokens identifying
-  the logon session. The utils setting will set the system to log non priority information
-  */
-
-  before('should initialize the service', function(callback) {
-    
-    this.timeout(20000);
-
-    try{
-      service.create({
-          secure:true,
-          encryptPayloads:true,
-          services:{
-            security:{
-              config:{
-                keyPair:{
-                  privateKey:'Kd9FQzddR7G6S9nJ/BK8vLF83AzOphW2lqDOQ/LjU4M=',
-                  publicKey:'AlHCtJlFthb359xOxR5kiBLJpfoC2ZLPLWYHN3+hdzf2'
-                }
-              }
-            }
-          }
-        }, 
-        function(e, happnInst){
-          if (e)
-            return callback(e);
-
-          happnInstance = happnInst;
-          callback();
-        });
-    }catch(e){
-      callback(e);
-    }
-  });
+  if (typeof window == 'undefined'){
+    var chai = require('chai')
+    , expect = chai.expect
+    , happn = require('../../lib/index')
+    , happn_client = happn.client
+  }
+  else{
+    expect = window.expect;
+    happn_client = window.HappnClient;
+  }
 
   after(function(done) {
-     socketClient.disconnect()
-    .then(eventEmitterClient.disconnect()
-    .then(happnInstance.stop()
-    .then(done)))
-    .catch(done);
+
+    if (socketClient){
+      socketClient.disconnect(done)
+    }
+    else done();
+    
   });
 
   var socketClient;
-  var eventEmitterClient;
-
-  var Crypto = require('consent-util-crypto');
-  var crypto = new Crypto();
-
-  var socketClientKeyPair = crypto.createKeyPair();
-  var eventEmitterClientKeyPair = crypto.createKeyPair();
+  var default_timeout = 10000;
 
   /*
     We are initializing 2 clients to test saving data against the database, one client will push data into the 
     database whilst another listens for changes.
   */
-  before('should initialize the clients', function(callback) {
+  before('should initialize the client', function(callback) {
       this.timeout(default_timeout);
 
       try {
@@ -77,33 +35,17 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
           config:{
             username:'_ADMIN',
             password:'happn',
-            keyPair:socketClientKeyPair
+            keyPair:{
+              publicKey:'AjN7wyfbEdI2LzWyFo6n31hvOrlYvkeHad9xGqOXTm1K', 
+              privateKey:'y5RTfdnn21OvbQrnBMiKBP9DURduo0aijMIGyLJFuJQ='
+            }
           }
-        },function(e, instance) {
+        }, function(e, instance) {
 
           if (e) return callback(e);
 
           socketClient = instance;
-
-          happn.client.create({
-            config:{
-              username:'_ADMIN', 
-              password:'happn',
-              keyPair:eventEmitterClientKeyPair
-            },
-            plugin: happn.client_plugins.intra_process,
-            context: happnInstance,
-            secure:true
-          })
-
-          .then(function(clientInstance){
-            eventEmitterClient = instance;
-            callback();
-          })
-
-          .catch(function(e){
-            callback(e);
-          });
+          callback();
 
         });
       } catch (e) {
@@ -121,9 +63,9 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
     try{
 
       //first listen for the change
-      eventEmitterClient.on('/e2e_test1/testsubscribe/data/event', {event_type:'set', count:1}, function(message){
+      socketClient.on('/e2e_test1/testsubscribe/data/event', {event_type:'set', count:1}, function(message){
 
-        expect(eventEmitterClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.be(0);
+        expect(socketClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.equal(0);
         callback();
 
       }, function(e){
@@ -132,7 +74,7 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
 
         if (!e){
 
-          expect(eventEmitterClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.be(1);
+          expect(socketClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.equal(1);
           //////////////////console.log('on subscribed, about to publish');
 
           //then make the change
@@ -157,7 +99,7 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
       //first listen for the change
       socketClient.on('/e2e_test1/testsubscribe/data/event', {event_type:'set', count:1}, function(message){
 
-        expect(socketClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.be(0);
+        expect(socketClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.equal(0);
         callback();
 
       }, function(e){
@@ -166,11 +108,11 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
 
         if (!e){
 
-          expect(socketClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.be(1);
+          expect(socketClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.equal(1);
           //////////////////console.log('on subscribed, about to publish');
 
           //then make the change
-          eventEmitterClient.set('/e2e_test1/testsubscribe/data/event', {property1:'property1',property2:'property2',property3:'property3'}, null, function(e, result){
+          socketClient.set('/e2e_test1/testsubscribe/data/event', {property1:'property1',property2:'property2',property3:'property3'}, null, function(e, result){
             ////////////////////////////console.log('put happened - listening for result');
           });
 
@@ -188,7 +130,7 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
     this.timeout(default_timeout);
 
     try{
-      var test_path_end = require('shortid').generate();
+      var test_path_end = '1';
 
       socketClient.set('e2e_test1/testsubscribe/data/' + test_path_end, {property1:'property1',property2:'property2',property3:'property3'}, {noPublish:true}, function(e, result){
       
@@ -200,10 +142,8 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
             ////////////console.log('new data results');
             ////////////console.log([e, results]);
 
-            expect(results.property1 == 'property1').to.be(true);
-
-            if (mode != 'embedded')
-              expect(results.created == results.modified).to.be(true);
+            expect(results.property1 == 'property1').to.equal(true);
+            expect(results.created == results.modified).to.equal(true);
 
             callback(e);
           });
@@ -222,7 +162,7 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
 
     try{
 
-      var test_path_end = require('shortid').generate();
+      var test_path_end = '2';
 
       socketClient.set('e2e_test1/testsubscribe/data/merge/' + test_path_end, {property1:'property1',property2:'property2',property3:'property3'}, null, function(e, result){
       
@@ -248,8 +188,8 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
             //////////////console.log('merge get results');
             //////////////console.log(results);
 
-            expect(results.property4).to.be('property4');
-            expect(results.property1).to.be('property1');
+            expect(results.property4).to.equal('property4');
+            expect(results.property1).to.equal('property1');
             
             callback();
 
@@ -268,7 +208,7 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
 
     //////////////////////////console.log('DOING COMPLEX SEARCH');
 
-    var test_path_end = require('shortid').generate();
+    var test_path_end = '3';
 
     var complex_obj = {
       regions:['North','South'],
@@ -297,22 +237,22 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
       limit:2}
 
     socketClient.set('/e2e_test1/testsubscribe/data/complex/' + test_path_end, complex_obj, null, function(e, put_result){
-      expect(e == null).to.be(true);
+      expect(e == null).to.equal(true);
       socketClient.set('/e2e_test1/testsubscribe/data/complex/' + test_path_end + '/1', complex_obj, null, function(e, put_result){
-        expect(e == null).to.be(true);
+        expect(e == null).to.equal(true);
 
         ////////////console.log('searching');
         socketClient.get('/e2e_test1/testsubscribe/data/complex*', {criteria:criteria1, options:options1}, function(e, search_result){
 
           ////////////console.log([e, search_result]);
 
-          expect(e == null).to.be(true);
-          expect(search_result.length == 1).to.be(true);
+          expect(e == null).to.equal(true);
+          expect(search_result.length == 1).to.equal(true);
 
           socketClient.get('/e2e_test1/testsubscribe/data/complex*', {criteria:criteria2, options:options2}, function(e, search_result){
 
-            expect(e == null).to.be(true);
-            expect(search_result.length == 2).to.be(true);
+            expect(e == null).to.equal(true);
+            expect(search_result.length == 2).to.equal(true);
 
             callback(e);
           });
@@ -337,8 +277,8 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
         //We perform the actual delete
         socketClient.remove('/e2e_test1/testsubscribe/data/delete_me', {noPublish:true}, function(e, result){
 
-          expect(e).to.be(null);
-          expect(result._meta.status).to.be('ok');
+          expect(e).to.equal(null);
+          expect(result._meta.status).to.equal('ok');
 
           ////////////////////console.log('DELETE RESULT');
           ////////////////////console.log(result);
@@ -359,16 +299,16 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
     this.timeout(default_timeout);
 
     try{
-      var test_path_end = require('shortid').generate();
+      var test_path_end = '4';
 
       socketClient.set('e2e_test1/testsubscribe/data/' + test_path_end, {property1:'property1',property2:'property2',property3:'property3'}, {noPublish:true}, function(e, insertResult){
       
-        expect(e).to.be(null);
+        expect(e).to.equal(null);
 
         socketClient.set('e2e_test1/testsubscribe/data/' + test_path_end, {property1:'property1',property2:'property2',property3:'property3', property4:'property4'}, {noPublish:true}, function(e, updateResult){
 
-          expect(e).to.be(null);
-          expect(updateResult._id == insertResult._id).to.be(true);
+          expect(e).to.equal(null);
+          expect(updateResult._id == insertResult._id).to.equal(true);
           callback();
 
         });
@@ -388,7 +328,7 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
     this.timeout(default_timeout);
 
     try{
-      var test_path_end = require('shortid').generate();
+      var test_path_end = '5';
 
       socketClient.set('e2e_test1/testsubscribe/data/' + test_path_end, {property1:'property1',property2:'property2',property3:'property3'}, null, function(e, result){
       
@@ -397,10 +337,10 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
             ////////////////////////console.log('new data results');
             ////////////////////////console.log(results);
 
-            expect(results.property1 == 'property1').to.be(true);
+            expect(results.property1 == 'property1').to.equal(true);
 
             // if (mode != 'embedded')
-            //  expect(results.payload[0].created == results.payload[0].modified).to.be(true);
+            //  expect(results.payload[0].created == results.payload[0].modified).to.equal(true);
 
             callback(e);
           });
@@ -420,16 +360,16 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
     this.timeout(default_timeout);
 
     try{
-      var test_path_end = require('shortid').generate();
+      var test_path_end = '6';
 
       socketClient.set('e2e_test1/testsubscribe/data/' + test_path_end, {property1:'property1',property2:'property2',property3:'property3'}, null, function(e, insertResult){
       
-        expect(e == null).to.be(true);
+        expect(e == null).to.equal(true);
 
         socketClient.set('e2e_test1/testsubscribe/data/' + test_path_end, {property1:'property1',property2:'property2',property3:'property3', property4:'property4'}, null, function(e, updateResult){
 
-          expect(e == null).to.be(true);
-          expect(updateResult._id == insertResult._id).to.be(true);
+          expect(e == null).to.equal(true);
+          expect(updateResult._id == insertResult._id).to.equal(true);
           callback();
 
         });
@@ -451,20 +391,20 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
 
     try{
 
-      var test_path_end = require('shortid').generate();  
+      var test_path_end = '7';  
 
       socketClient.setSibling('e2e_test1/siblings/' + test_path_end, {property1:'sib_post_property1',property2:'sib_post_property2'}, function(e, results){
 
-        expect(e == null).to.be(true);
+        expect(e == null).to.equal(true);
 
         socketClient.setSibling('e2e_test1/siblings/' + test_path_end, {property1:'sib_post_property1',property2:'sib_post_property2'}, function(e, results){
 
-          expect(e == null).to.be(true);
+          expect(e == null).to.equal(true);
 
           //the child method returns a child in the collection with a specified id
           socketClient.get('e2e_test1/siblings/' + test_path_end + '/*', null, function(e, getresults){
-            expect(e == null).to.be(true);
-            expect(getresults.length == 2).to.be(true);
+            expect(e == null).to.equal(true);
+            expect(getresults.length == 2).to.equal(true);
             callback(e);
           });
         });
@@ -486,16 +426,16 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
     try{
 
       //first listen for the change
-      eventEmitterClient.on('/e2e_test1/testsubscribe/data/event', {event_type:'set', count:1}, function(message){
+      socketClient.on('/e2e_test1/testsubscribe/data/event', {event_type:'set', count:1}, function(message){
 
-        expect(eventEmitterClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.be(0);
+        expect(socketClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.equal(0);
         callback();
 
       }, function(e){
 
         if (!e){
 
-          expect(eventEmitterClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.be(1);
+          expect(socketClient.events['/SET@/e2e_test1/testsubscribe/data/event'].length).to.equal(1);
 
           ////////////////////////////console.log('on subscribed, about to publish');
 
@@ -516,20 +456,20 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
 
   it('should get using a wildcard', function(callback) {
 
-    var test_path_end = require('shortid').generate();
+    var test_path_end = '8';
 
     socketClient.set('e2e_test1/testwildcard/' + test_path_end, {property1:'property1',property2:'property2',property3:'property3'}, null, function(e, insertResult){
-      expect(e == null).to.be(true);
+      expect(e == null).to.equal(true);
       socketClient.set('e2e_test1/testwildcard/' + test_path_end + '/1', {property1:'property1',property2:'property2',property3:'property3'}, null, function(e, insertResult){
-        expect(e == null).to.be(true);
+        expect(e == null).to.equal(true);
       
         socketClient.get('e2e_test1/testwildcard/' + test_path_end + '*', null, function(e, results){
           
-          expect(results.length == 2).to.be(true);
+          expect(results.length == 2).to.equal(true);
 
           socketClient.getPaths('e2e_test1/testwildcard/' + test_path_end + '*', function(e, results){
 
-            expect(results.length == 2).to.be(true);
+            expect(results.length == 2).to.equal(true);
             callback(e);
 
           });
@@ -550,21 +490,21 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
           //////////////////console.log('did delete set');
           //path, event_type, count, handler, done
           //We listen for the DELETE event
-          eventEmitterClient.on('/e2e_test1/testsubscribe/data/delete_me', {event_type:'remove', count:1}, function(eventData){
+          socketClient.on('/e2e_test1/testsubscribe/data/delete_me', {event_type:'remove', count:1}, function(eventData){
 
             ////console.log('on count 1 delete ');
             //////////////////console.log(message);
 
             //we are looking at the event internals on the listener to ensure our event management is working - because we are only listening for 1
             //instance of this event - the event listener should have been removed 
-            ////console.log('eventEmitterClient.events');
-            ////console.log(eventEmitterClient.events);
-            expect(eventEmitterClient.events['/REMOVE@/e2e_test1/testsubscribe/data/delete_me'].length).to.be(0);
+            ////console.log('socketClient.events');
+            ////console.log(socketClient.events);
+            expect(socketClient.events['/REMOVE@/e2e_test1/testsubscribe/data/delete_me'].length).to.equal(0);
 
             ////console.log(eventData);
 
             //we needed to have removed a single item
-            expect(eventData.removed).to.be(1);
+            expect(eventData.removed).to.equal(1);
 
             ////////////////////////////console.log(message);
 
@@ -575,9 +515,9 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
             ////////////console.log('ON HAS HAPPENED: ' + e);
 
             if (!e){
-              ////console.log('eventEmitterClient.events, pre');
-              ////console.log(eventEmitterClient.events);
-              expect(eventEmitterClient.events['/REMOVE@/e2e_test1/testsubscribe/data/delete_me'].length).to.be(1);
+              ////console.log('socketClient.events, pre');
+              ////console.log(socketClient.events);
+              expect(socketClient.events['/REMOVE@/e2e_test1/testsubscribe/data/delete_me'].length).to.equal(1);
 
               //////////////////console.log('subscribed, about to delete');
 
@@ -608,22 +548,22 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
 
     var currentListenerId;
 
-    eventEmitterClient.on('/e2e_test1/testsubscribe/data/on_off_test', {event_type:'set', count:0}, function(message){
+    socketClient.on('/e2e_test1/testsubscribe/data/on_off_test', {event_type:'set', count:0}, function(message){
 
       //we detach all listeners from the path here
       ////console.log('ABOUT OFF PATH');
-      eventEmitterClient.off('/e2e_test1/testsubscribe/data/on_off_test', function(e){
+      socketClient.off('/e2e_test1/testsubscribe/data/on_off_test', function(e){
 
         if (e)
           return callback(new Error(e));
 
-        eventEmitterClient.on('/e2e_test1/testsubscribe/data/on_off_test', {event_type:'set', count:0}, 
+        socketClient.on('/e2e_test1/testsubscribe/data/on_off_test', {event_type:'set', count:0}, 
         function(message){
 
           ////console.log('ON RAN');
           ////console.log(message);
 
-          eventEmitterClient.off(currentListenerId, function(e){
+          socketClient.off(currentListenerId, function(e){
 
             if (e)
               return callback(new Error(e));
@@ -667,7 +607,7 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
 
     this.timeout(10000);
     
-    eventEmitterClient.onAll(function(eventData, meta){
+    socketClient.onAll(function(eventData, meta){
 
       if (meta.action == '/REMOVE@/e2e_test1/testsubscribe/data/catch_all' || 
               meta.action == '/SET@/e2e_test1/testsubscribe/data/catch_all')
@@ -699,7 +639,7 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
 
     var onHappened = false;
 
-    eventEmitterClient.onAll(function(message){
+    socketClient.onAll(function(message){
 
       onHappened = true;
       callback(new Error('this wasnt meant to happen'));
@@ -708,7 +648,7 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
 
       if (e) return callback(e);
       
-      eventEmitterClient.on('/e2e_test1/testsubscribe/data/off_all_test', {event_type:'set', count:0}, 
+      socketClient.on('/e2e_test1/testsubscribe/data/off_all_test', {event_type:'set', count:0}, 
         function(message){
           onHappened = true;
           callback(new Error('this wasnt meant to happen'));
@@ -716,7 +656,7 @@ describe('c2_websockets_embedded_sanity_encryptedpayloads', function() {
         function(e){
           if (e) return callback(e);
 
-          eventEmitterClient.offAll(function(e){
+          socketClient.offAll(function(e){
             if (e) return callback(e);
 
             socketClient.set('/e2e_test1/testsubscribe/data/off_all_test', {property1:'property1',property2:'property2',property3:'property3'}, null, function(e, put_result){